+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 3.10.1
 * Remove Gutter Logging
+* ADD `defaultHeaderHeight` defaulting to 0 in arcane theme
 
 ## 3.10.0
 * BREAKING: Removed the context parameter for `MenuButton` `onPressed`. If you need the menu button context specifically, use `onContextPressed` in the `MenuButton` instead. Both callbacks will fire if both are defined.